Refactor StreamCatalogService to get rid of dependencies from stream-catalog
* introduce 'streamline-actions' module which represents 'topology actions'
  * move relevant classes into actions
  * also introduce 'streamline-actions-storm' for Storm runner
* introduce TopologyActionsService and TopologyMetricsService
  * move StreamCatalogService methods to relevant classes
* get rid of dependency for actions and metrics from stream-catalog
  * now 'actions' and 'metrics' can depend on stream-catalog
