+++ pom.xml
@@ -747,6 +747,11 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.streamline</groupId>
+                <artifactId>streamline-actions</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.streamline</groupId>
                 <artifactId>streamline-metrics-storm</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -796,6 +801,11 @@
             <dependency>
                 <groupId>org.apache.streamline</groupId>
                 <artifactId>streamline-runtime-storm</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.streamline</groupId>
+                <artifactId>streamline-actions-storm</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
