+++ streams/runners/storm/actions/pom.xml
@@ -3,17 +3,18 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>streams</artifactId>
+        <artifactId>streamline-runners-storm</artifactId>
         <groupId>org.apache.streamline</groupId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>streamline-metrics</artifactId>
+    <artifactId>streamline-actions-storm</artifactId>
+
     <dependencies>
         <dependency>
             <groupId>org.apache.streamline</groupId>
-            <artifactId>common</artifactId>
+            <artifactId>streamline-actions</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.streamline</groupId>
@@ -21,7 +22,8 @@
         </dependency>
         <dependency>
             <groupId>org.apache.streamline</groupId>
-            <artifactId>streamline-catalog</artifactId>
+            <artifactId>streamline-layout-storm</artifactId>
         </dependency>
     </dependencies>
+
 </project>
