+++ streams/runners/storm/pom.xml
@@ -15,6 +15,7 @@
         <module>layout</module>
         <module>runtime</module>
         <module>metrics</module>
+        <module>actions</module>
     </modules>
